+++ manual/conf.py
@@ -26,9 +26,9 @@ copyright = '2014-2022, Brandon Castellano'
 author = 'Brandon Castellano'
 
 # The short X.Y version
-version = '0.6'
+version = '0.6.1-dev'
 # The full version, including alpha/beta/rc tags
-release = 'v0.6'
+release = 'v0.6.1-dev'
 
 
 # -- General configuration ---------------------------------------------------